+++ xc/lib/Xtst/XTest.c
@@ -32,6 +32,7 @@ other dealings in this Software without prior written authorization
 from the X Consortium.
 
 */
+/* $XFree86: xc/lib/Xtst/XTest.c,v 1.1.1.1.12.2 1998/10/04 15:22:55 hohndel Exp $ */
 
 #define NEED_REPLIES
 #include "Xlibint.h"
@@ -82,7 +83,7 @@ get_xinput_base(dpy)
     first_event = 0;
 
     XQueryExtension(dpy, INAME, &major_opcode, &first_event, &first_error);
-    return (XPointer)first_event;
+    return (XPointer)(long)first_event;
 }
 
 static XEXT_GENERATE_FIND_DISPLAY (find_display, xtest_info,
@@ -283,7 +284,7 @@ send_axes(dpy, info, req, dev, first_axis, axes, n_axes)
 
     req->deviceid |= MORE_EVENTS;
     req->length += ((n_axes + 5) / 6) * (SIZEOF(xEvent) >> 2);
-    ev.type = XI_DeviceValuator + (int)info->data;
+    ev.type = XI_DeviceValuator + (long)info->data;
     ev.deviceid = dev->device_id;
     ev.num_valuators = n_axes;
     ev.first_valuator = first_axis;
@@ -331,7 +332,7 @@ XTestFakeDeviceKeyEvent(dpy, dev, keycode, is_press, axes, n_axes, delay)
     req->reqType = info->codes->major_opcode;
     req->xtReqType = X_XTestFakeInput;
     req->type = is_press ? XI_DeviceKeyPress : XI_DeviceKeyRelease;
-    req->type += (int)info->data;
+    req->type += (int)(long)info->data;
     req->detail = keycode;
     req->time = delay;
     req->deviceid = dev->device_id;
@@ -361,7 +362,7 @@ XTestFakeDeviceButtonEvent(dpy, dev, button, is_press, axes, n_axes, delay)
     req->reqType = info->codes->major_opcode;
     req->xtReqType = X_XTestFakeInput;
     req->type = is_press ? XI_DeviceButtonPress : XI_DeviceButtonRelease;
-    req->type += (int)info->data;
+    req->type += (int)(long)info->data;
     req->detail = button;
     req->time = delay;
     req->deviceid = dev->device_id;
@@ -390,7 +391,7 @@ XTestFakeProximityEvent(dpy, dev, in_prox, axes, n_axes, delay)
     req->reqType = info->codes->major_opcode;
     req->xtReqType = X_XTestFakeInput;
     req->type = in_prox ? XI_ProximityIn : XI_ProximityOut;
-    req->type += (int)info->data;
+    req->type += (int)(long)info->data;
     req->time = delay;
     req->deviceid = dev->device_id;
     if (n_axes)
@@ -419,7 +420,7 @@ XTestFakeDeviceMotionEvent(dpy, dev, is_relative,
     GetReq(XTestFakeInput, req);
     req->reqType = info->codes->major_opcode;
     req->xtReqType = X_XTestFakeInput;
-    req->type = XI_DeviceMotionNotify + (int)info->data;
+    req->type = XI_DeviceMotionNotify + (int)(long)info->data;
     req->detail = is_relative;
     req->time = delay;
     req->deviceid = dev->device_id;
