+++ xc/programs/Xserver/Xprint/ps/PsWindow.c
@@ -48,7 +48,7 @@
 **    *********************************************************
 ** 
 ********************************************************************/
-/* $XFree86: xc/programs/Xserver/Xprint/ps/PsWindow.c,v 1.6 1997/01/14 22:16:19 dawes Exp $ */
+/* $XFree86: xc/programs/Xserver/Xprint/ps/PsWindow.c,v 1.6.2.1 1998/10/04 13:36:45 hohndel Exp $ */
 
 #include <stdio.h>
 #include <string.h>
@@ -354,14 +354,14 @@ PsPaintWindow(
         }
         break;
       case GCClipMask:
-        if( (pointer)pGC->clientClipType!=(pointer)CT_NONE )
+        if( (pointer)(long)pGC->clientClipType!=(pointer)CT_NONE )
         {
           gcmask |= index;
           gcval[i++] = (pointer)CT_NONE;
         }
         break;
       case GCSubwindowMode:
-        if( (pointer)pGC->subWindowMode!=newValues[SUBWINDOW] )
+        if( (pointer)(long)pGC->subWindowMode!=newValues[SUBWINDOW] )
         {
           gcmask |= index;
           gcval[i++] = newValues[SUBWINDOW];
@@ -375,7 +375,7 @@ PsPaintWindow(
         }
         break;
       case GCFillStyle:
-        if( (pointer)pGC->fillStyle!=newValues[FILLSTYLE] )
+        if( (pointer)(long)pGC->fillStyle!=newValues[FILLSTYLE] )
         {
           gcmask |= index;
           gcval[i++] = newValues[FILLSTYLE];
