+++ xc/programs/Xserver/record/set.c
@@ -29,6 +29,7 @@ other dealings in this Software without prior written authorization
 from the X Consortium.
 
 */
+/* $XFree86: xc/programs/Xserver/record/set.c,v 1.1.1.1.4.2 1998/10/04 15:23:05 hohndel Exp $ */
 
 /*
 
@@ -199,7 +200,7 @@ BitVectorIterateSet(pSet, pIter, pInterval)
     RecordSetIteratePtr pIter;
     RecordSetInterval *pInterval;
 {
-    int iterbit = (int)pIter;
+    int iterbit = (int)(long)pIter;
     int b;
 
     b = BitVectorFindBit(pSet, iterbit, TRUE);
@@ -208,7 +209,7 @@ BitVectorIterateSet(pSet, pIter, pInterval)
 
     b = BitVectorFindBit(pSet, b, FALSE);
     pInterval->last = (b < 0) ? ((BitVectorSetPtr)pSet)->maxMember : b - 1;
-    return (RecordSetIteratePtr)(pInterval->last + 1);
+    return (RecordSetIteratePtr)(long)(pInterval->last + 1);
 }
 
 RecordSetOperations BitVectorSetOperations = {
@@ -517,7 +518,7 @@ RecordCreateSet(pIntervals, nIntervals, pMem, memsize)
 					&pCreateSet);
     if (pMem)
     {
-	if ( ((int)pMem & (alignment-1) ) || memsize < size)
+	if ( ((long)pMem & (alignment-1) ) || memsize < size)
 	    return NULL;
     }
     return (*pCreateSet)(pIntervals, nIntervals, pMem, size);
